decomp progress: 11 functions to 100% across SDK/system/band3
100%:
- ClearArena 98.5% → 100% (inline OSIsMEM1Region)
- ClearMEM2Arena 98.1% → 100% (inline OSIsMEM2Region)
- __OSInitAudioSystem 97.7% → 100% (OS_TICKS_DELTA inline)
- ConfigureVideo (OSFatal) 99.8% → 100% (NON_INT scan mode)
- AppLabel::SetSongNameWithNumber 99.1% → 100% (cc ternary)
- PassiveMessenger::HasMessages 99.8% → 100% ((int) cast)
- WiiEnviron::SetLight 98.2% → 100% (int id)
- VorbisReader::ConsumeData 98.5% → 100% (volatile cast)
- RockCentral::ConvertToStr 98.7% → 100% (unsigned char + mask)
- RockCentral::SendMsg 96.4% → 100% (int* cast)
- SetlistMergePanel::UpdateProgressInfo 98.1% → 99.9% (Abs ternary)
- CalibrationPanel::UpdateAnimation 98.7% → 99.9% (% 2)

Partial:
- CharProvider::AddCharactersFromProfile 98.7% → 98.8%
- Spotlight::Poll 97.1% → 97.7% (combined early returns)

+++ src/band3/meta_band/PassiveMessenger.cpp
@@ -357,7 +357,7 @@ bool PassiveMessenger::HasMessages() const {
         if (slot) {
             PassiveMessageQueue *pQueue = slot->GetMessageQueue();
             MILO_ASSERT(pQueue, 0x299);
-            if (pQueue->mQueue.size() > 0)
+            if ((int)pQueue->mQueue.size() > 0)
                 return true;
             if (pQueue->mTimer.Running())
                 return true;
